// src/app/lib/generated/models/GetChatbotListResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ListChatbotItem } from './ListChatbotItem';
import {
    ListChatbotItemFromJSON,
    ListChatbotItemFromJSONTyped,
    ListChatbotItemToJSON,
    ListChatbotItemToJSONTyped,
} from './ListChatbotItem';

/**
 * 
 * @export
 * @interface GetChatbotListResponse
 */
export interface GetChatbotListResponse {
    /**
     * 
     * @type {Array<ListChatbotItem>}
     * @memberof GetChatbotListResponse
     */
    girls: Array<ListChatbotItem>;
    /**
     * 
     * @type {Array<ListChatbotItem>}
     * @memberof GetChatbotListResponse
     */
    anime: Array<ListChatbotItem>;
}

/**
 * Check if a given object implements the GetChatbotListResponse interface.
 */
export function instanceOfGetChatbotListResponse(value: object): value is GetChatbotListResponse {
    if (!('girls' in value) || value['girls'] === undefined) return false;
    if (!('anime' in value) || value['anime'] === undefined) return false;
    return true;
}

export function GetChatbotListResponseFromJSON(json: any): GetChatbotListResponse {
    return GetChatbotListResponseFromJSONTyped(json, false);
}

export function GetChatbotListResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetChatbotListResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'girls': ((json['girls'] as Array<any>).map(ListChatbotItemFromJSON)),
        'anime': ((json['anime'] as Array<any>).map(ListChatbotItemFromJSON)),
    };
}

export function GetChatbotListResponseToJSON(json: any): GetChatbotListResponse {
    return GetChatbotListResponseToJSONTyped(json, false);
}

export function GetChatbotListResponseToJSONTyped(value?: GetChatbotListResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'girls': ((value['girls'] as Array<any>).map(ListChatbotItemToJSON)),
        'anime': ((value['anime'] as Array<any>).map(ListChatbotItemToJSON)),
    };
}